v0.2.0 in progess, added bnv
