+++ app/src/main/java/com/eastmonster/accountbook/AccountRepository.java
@@ -5,7 +5,6 @@ import android.app.Application;
 import androidx.lifecycle.LiveData;
 
 import java.util.List;
-import java.util.Vector;
 
 public class AccountRepository {
     private AccountDao mAccountDao;
