+++ app/src/main/java/com/eastmonster/accountbook/NewAccountActivity.java
@@ -1,11 +1,9 @@
 package com.eastmonster.accountbook;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.ViewModelProvider;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
