+++ app/src/main/java/com/eastmonster/accountbook/StatListAdapter.java
@@ -3,43 +3,10 @@ package com.eastmonster.accountbook;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.DiffUtil;
-import androidx.recyclerview.widget.ListAdapter;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.List;
 
-
-class StatItem implements Comparable<StatItem>{
-    private int type;
-    private double sumAmount;
-
-    public StatItem(int type, double sumAmount) {
-        this.type = type;
-        this.sumAmount = sumAmount;
-    }
-
-    public int getType() {
-        return type;
-    }
-
-    public void setType(int type) {
-        this.type = type;
-    }
-
-    public double getSumAmount() {
-        return sumAmount;
-    }
-
-    public void setSumAmount(double sumAmount) {
-        this.sumAmount = sumAmount;
-    }
-
-    @Override
-    public int compareTo(StatItem s) { // 降序显示
-        return this.sumAmount >= s.getSumAmount() ? 1 : 0;
-    }
-}
 
 public class StatListAdapter extends RecyclerView.Adapter<StatViewHolder> {
