+++ app/src/main/res/drawable-anydpi/ic_statistics.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:tint="#333333"
+    android:alpha="0.6">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M22,22H2v-2h20V22zM10,2H7v16h3V2zM17,8h-3v10h3V8z"/>
+</vector>
